% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweet_extensions.R
\name{get_api_request_limits}
\alias{get_api_request_limits}
\title{Get Twitter API request limits}
\usage{
get_api_request_limits(query, token = rtweet::get_token())
}
\arguments{
\item{query}{query string (e.g., 'statuses/user_timeline')}

\item{token}{an Twitter OAuth \code{rtweet}'s 'Token' object.
Defaults to \code{rtweet}'s \code{\link[rtweet]{get_token}}}
}
\value{
Integer value giving the number of remaining available request for a given query.
    (Vector is named by query in order to allow for name-based indexing)
}
\description{
Get Twitter API request limits
}
