% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_required_date_ranges}
\alias{get_required_date_ranges}
\title{Get required date ranges to be requested}
\usage{
get_required_date_ranges(.dates, .since, .until)
}
\arguments{
\item{.dates}{a vector of dates or \code{NULL}, defining the (start and/or end)
dates taken as given.
Alternatively, \code{.dates} can be a vector of integer values that will be converted
to dates using \code{\link[lubridate]{as_date}}
(hence, \code{\link[lubridate]{origin}}, that is, '1970-01-01 UTC' will be used as origin date).}

\item{.since}{a date (format '\%Y-\%m-\%d'), specifying the start of the date range to be requested}

\item{.until}{a date (format '\%Y-\%m-\%d'), specifying the end of the date range to be requested}
}
\value{
Either of the following two
\enumerate{
   \item A list of date range start and end dates in the date range spaned from \code{.since} to \code{.until}
       that is \bold{not} covered by dates in \code{.dates}.
   \item \code{NULL} if length of \code{dates} is zero,
       or the requested time range is convered by the date range spaned from \code{.since} to \code{.until}.
}
}
\description{
Function takes a vector of dates, a start and end date, and returns a list of adjacent date sequences that are \bold{not} covered by the vector of dates.
}
\details{
This function is an internal helper. All dates need to be in \code{YYYY-mm-dd} ('\%Y-\%m-\%d') format.
}
\examples{
dates <- seq(Sys.Date()-30, Sys.Date()-5, by = "day")
since <- Sys.Date() -35
until <- Sys.Date()

# returns list with one date range vector (spanning since - until)
twscrape:::get_required_date_ranges(NULL, since, until)

# returns list with two date range vectors
twscrape:::get_required_date_ranges(Sys.Date()-1, since, until)

# returns list with two date range vectors
twscrape:::get_required_date_ranges(dates, since, until)
}
