% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_running_dockers}
\alias{get_running_dockers}
\title{Get running docker containers}
\usage{

  get_running_dockers(format.string = "table {{.Names}},{{.Image}},{{.ID}},{{.Ports}}")
}
\arguments{
\item{format.string}{the formatting string passed to \code{`docker ps --format '<format string>'`}.
Defaults to "table {{.Names}},{{.Image}},{{.ID}},{{.Ports}}"}
}
\value{
A data frame with columns as defined by \code{format.string},
     i.e. 'names', 'image', 'container_id', and 'ports' for the default formatting string.
     Each row representes one running container (hence, 0 rows if no container is running).
}
\description{
Functions uses \code{\link[base]{system}} to see what docker containers are running and
    returns running containers as a data frame.
    It throws an \code{Error} if docker is not installed or docker deamon is not running (incl informative message).
}
\details{
The output is generated by runing \code{`docker ps --format '...'`} in the shell.
    Hence, a function call throws an \code{Error} if either
        (i) operating system is Windows,
        (ii) docker is not installed on system, or
        (iii) the docker deamon is not running.
}
\section{Warning}{

    Currently implemented for Mac OSX/Unix/Linux only.
}

