% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_scrape_tweets.R
\name{get_tweet_ids_meta}
\alias{get_tweet_ids_meta}
\title{Get tweet IDs meta data from tweet IDs JSON file}
\usage{
get_tweet_ids_meta(file, .uid.pattern = "^tw_user_(\\\\d+)_.+",
  .daterange.pattern = "^[[:alnum:]_]+_tweet_ids_([[:alnum:]_-]+)\\\\.json$")
}
\arguments{
\item{file}{Path to the tweets IDs JSON file}

\item{.uid.pattern}{PERL compatible regular expression used to parse numeric user ID from \code{file} basename}

\item{.daterange.pattern}{PERL compatible regular expression used to parse date range from \code{file} basename}
}
\value{
a \link[tibble]{tibble} data frame with columns
\itemize{
    \item 'user_id' (parsed from JSON file stem using \code{.uid.pattern}),
    \item 'screen_name' (parent element of JSON file),
    \item 'start' (first date parsed from \code{.daterange.pattern}),
    \item 'end' (second date parsed from \code{.daterange.pattern}),
    \item 'since' (Twitter search query'since' date parsed from JSON elements),
    \item 'until' (Twitter search query 'until' date parsed from JSON elements),
    \item 'n_tweet_ids' (length of list in nested 'ids' elements of JSON elements).
    \item 'file' (as passed to \code{file}).
}
}
\description{
Returns number of IDs collected per account and date range window
}
