% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_scrape_tweets.R
\name{get_user_tweet_ids_in_date_range}
\alias{get_user_tweet_ids_in_date_range}
\title{Get a user's tweet IDs for a given date range}
\usage{
get_user_tweet_ids_in_date_range(screen.name, user.id, remdr, since, until,
  date.interval = "year", sleep = 2, .write.out = TRUE, .data.path,
  .file.stem = paste0("tw_user_", user.id, "_tweet_ids_\%s"),
  verbose = TRUE)
}
\arguments{
\item{screen.name}{is the screen (or account) name of a twitter user}

\item{user.id}{is the user ID of a twitter user}

\item{remdr}{an \bold{active} RSelenium \code{\link[RSelenium]{remoteDriver}} object 
(check \code{remdr$getStatus()} to see if the driver is running.)}

\item{since}{a date (format '\%Y-\%m-\%d'), specifying the start of the date range to be requested}

\item{until}{a date (format '\%Y-\%m-\%d'), specifying the end of the date range to be requested}

\item{date.interval}{date interval passed to 'by' argument of \code{\link[base]{seq.Date}}. Defaults to 'year'.}

\item{sleep}{Seconds to pause between non-adjacent date ranges. Defautls to 2.}

\item{.write.out}{logical. write out tweet IDs as JSON to disk?
If \code{TRUE} (the default), JSON file will be written 
to path \code{.data.path}}

\item{.data.path}{Path to look at for existing tweet ID files
Also the path where new ID files are written if \code{write.out = TRUE}.}

\item{.file.stem}{JSON file name stem (stem ignores date ranges)
Defaults to glob 'tw_user_<\code{tw.account}>_tweet_ids_*.json'.
Used both for looking for existing tweet ID JSON files, and to name new ones when writing to disk.}

\item{verbose}{logical. Print out status messages?}
}
\value{

}
\description{
Given a start and end date, function looks for tweet ID files already to disk
   and gets tweet IDs for the remaining date range(s) by calling \link{scrape_tweets}.
}
\section{WARNING}{

\itemize{
    \item Function presuposses an active remote Selenium driver.     
    \item Function only accepts dates in format '\%Y-\%m-\%d' (Year-month-day: 'YYYY-mm-dd')     
}
}

