% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_tweets.R
\name{get_user_tweets}
\alias{get_user_tweets}
\title{Get a user's tweets for a given date range}
\usage{
get_user_tweets(ids.file, user.id, since, until,
  token = rtweet::get_token(), .write.out = TRUE,
  .data.path = "/Users/licht/Dropbox/smdata/sink/candidates/at/tweets/",
  .file.stem = paste0("tw_user_", user.id, "_tweets_\%s.RData"),
  .write.fun = saveRDS, verbose = TRUE)
}
\arguments{
\item{ids.file}{path to a tweet IDs JSON (as written by \link{scrape_tweet_ids} or \link{get_user_tweet_ids})}

\item{user.id}{is the user ID of a twitter user}

\item{since}{a date (format '\%Y-\%m-\%d'), specifying the start of the date range to be requested}

\item{until}{a date (format '\%Y-\%m-\%d'), specifying the end of the date range to be requested}

\item{token}{an Twitter OAuth \code{rtweet}'s 'Token' object.
Defaults to \code{rtweet}'s \code{get_token}.}

\item{.write.out}{logical. write out tweet IDs as JSON to disk?
If \code{TRUE} (the default), JSON file will be written
to path \code{.data.path}}

\item{.data.path}{Path to look at for existing tweet ID files
Also the path where new ID files are written if \code{write.out = TRUE}.}

\item{.file.stem}{file name stem (stem ignores date ranges).
Defaults to glob 'tw_user_<\code{user.id}>_tweets_*.RData'.
Used both for looking for existing tweets data files, and to name new ones when writing to disk.}

\item{.write.fun}{function used to write data output list.
Note that you cannot use \link[utils]{write.csv} or any other table-like writer, as the outputted data is a list object.}

\item{verbose}{logical. Print out status messages?}
}
\value{
A list object with three elements
     \enumerate{
          \item 'tweets': \code{\link[tibble]{tibble}} data frame containing the tweets data.
          \item 'data_pathes': a chracter vector, specifiying the paths of \emph{all} tweets data files if any exist or new were written
          \item 'failures': \code{\link[tibble]{tibble}} data frame reporting warnings and errors raised while trying to get tweets data.
     }
}
\description{
Given a tweet IDs JSON, a user.id, and a start and end date,
    function looks for existing tweets data
    and gets tweets for the remaining date range(s) by calling \code{\link{get_all_tweets}}.
}
\section{WARNING}{

\itemize{
    \item Function presuposses a valid Twitter OAuth token.
    \item Function only accepts dates in format '\%Y-\%m-\%d' (Year-month-day: 'YYYY-mm-dd')
}
}

