% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweet_extensions.R
\name{read_tweets_data}
\alias{read_tweets_data}
\title{Read tweets data from disk}
\usage{
read_tweets_data(paths, fields = tw_status_cols(status.location = FALSE,
  user.bio = FALSE, user.engagements = FALSE, user.profile = FALSE,
  user.misc = FALSE), read.fun = readRDS)
}
\arguments{
\item{paths}{character vector specifiying the paths to twitter data files}

\item{fields}{names of columns (fields) to be extracted and row-binded in output data frame}

\item{read.fun}{reader function (needs to be compatible with formats of files passed to \code{paths})}
}
\value{
A \link[tibble]{tibble} data frame with columns as specified by argument \code{fields}
}
\description{
Given a vector of file paths, function returns a (row-stacked)
    data frame object of the contained tweets data
}
\details{
The files to be read need to be tweet data frame as returned by \code{rtweet}'s \link[rtweet]{lookup_statuses}.
    The file format needs to be readable by the reader function passed to parameter \code{read.fun}.
}
