% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweet_extensions.R
\name{tw_status_cols}
\alias{tw_status_cols}
\title{Get \code{rtweet}'s status-lookup columns}
\usage{
tw_status_cols(status = TRUE, user = TRUE, status.plus = status,
  status.engagements = status, status.entities = status,
  status.entities.hashtags = status.entities,
  status.entities.symbols = status.entities,
  status.entities.urls = status.entities,
  status.entities.media = status.entities, status.reply = status,
  status.quote = status, status.quote.status = status.quote,
  status.quote.user = status.quote, status.retweet = status,
  status.retweet.status = status.retweet,
  status.retweet.user = status.retweet, status.location = status,
  status.location.place = status.location,
  status.location.coords = status.location, user.bio = user,
  user.engagements = user, user.profile = user, user.misc = user)
}
\arguments{
\item{status}{column names of main status info (status ID and text, created time)}

\item{user}{column names of main user info (user ID, screen name and name, and account created time)}

\item{status.plus}{column names of status meta info}

\item{status.engagements}{column names of status engagement count fields (No. favorites, retweets, quotes, replies)}

\item{status.entities}{column names of status entities objects}

\item{status.entities.hashtags}{column names of status hashtags object}

\item{status.entities.symbols}{column names of status symbols object}

\item{status.entities.urls}{column names of status urls object}

\item{status.entities.media}{column names of status media object}

\item{status.reply}{column names of status reply object (user ID, status ID, created time)}

\item{status.quote}{column names of status quote object (user ID, status ID and text, created time)}

\item{status.quote.status}{column names of additional status-related fields of status quote object}

\item{status.quote.user}{column names of additional user-related fields of status quote object}

\item{status.retweet}{column names of status retweet object (user ID, status ID and text, created time)}

\item{status.retweet.status}{column names of additional status-related fields of status retweet object}

\item{status.retweet.user}{column names of additional user-related fields of status retweet object}

\item{status.location}{column names of status' location objects}

\item{status.location.place}{column names of status' place info}

\item{status.location.coords}{column names of status' coordination info}

\item{user.bio}{column names of user account-related fields}

\item{user.engagements}{column names of user engagement fields}

\item{user.profile}{column names of user profile-related fields}

\item{user.misc}{user column names of miscalaneus user info fields}
}
\value{
a character vector of column names compatible with the column names of
    the data frame obejct returned by valid calls to \code{rtweet}'s \link[rtweet]{lookup_statuses}
}
\description{
Function returns column names of fields in twitter object returned
    on Twitter API call via the \code{rtweet}'s \link[rtweet]{lookup_statuses}
}
