% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_replace_patterns.R
\docType{data}
\name{tweet_replace_patterns}
\alias{tweet_replace_patterns}
\title{Regex patterns of special tokens in tweet texts}
\format{A named character vector with five elements (names: :
\describe{
  \item{"@_": regex pattern capturing Twitter user mentions}
  \item{"#_": regex pattern capturing Twitter hastags}
  \item{"M_": regex pattern capturing email addresses (see \url{https://emailregex.com/})}
  \item{"U_": regex pattern capturing URLs (see \url{http://daringfireball.net/2010/07/improved_regex_for_matching_urls})}
  \item{"E_": regex pattern capturing common Emojis (thse listed here \url{http://www.unicode.org/emoji/charts/full-emoji-list.html#smileys_&_emotion})}
}}
\usage{
tweet_replace_patterns
}
\description{
A vector mapping replacment place holders to regex patterns matching
    special tokens in tweets, namely
      mentions (@...),
      hastags (#... ),
      email addresses,
      URLs, and
      common Emojis.
}
\keyword{datasets}
