% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_dates}
\alias{validate_dates}
\title{Validate dates}
\usage{
validate_dates(..., .format = "\%Y-\%m-\%d")
}
\arguments{
\item{...}{a comma-separated list of presumable date values}

\item{.format}{an ISO date fromat to be validated
Default is '\%Y-\%m-\%d'.}
}
\value{
A list, one element for each element passed to \code{...}.
    Elements NA when
}
\description{
Function validates format of dates
    throws warning if a given value has not the required date format
    or cannot be converted to a valid date
}
\examples{
\dontrun{
test_ <- validate_dates(1234, Sys.Date(), "2019-02-30")
test_
lapply(test_, class)
}
}
