% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wait_until}
\alias{wait_until}
\title{Wait until}
\usage{
wait_until(until.time, pause.secs = 5)
}
\arguments{
\item{until.time}{POSIXlt time stamp}

\item{pause.secs}{integer. seconds to pause between checking if \code{until.time} has been reached}
}
\value{
\code{NULL}
}
\description{
Wait until
}
