% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_scrape_tweets.R
\name{scrape_tweet_ids}
\alias{scrape_tweet_ids}
\title{Scrape Tweets IDs from screen}
\usage{
scrape_tweet_ids(tw.account, remdr, since.date, until.date,
  date.interval = "month", write.out = TRUE, write.out.path,
  write.out.name = sprintf("tw_user_\%s_tweet_ids_\%s.json", tw.account,
  paste0(since.date, "_to_", until.date)), sleep = 0.5, verbose = TRUE)
}
\arguments{
\item{tw.account}{a scalar character vector, specifying a Twitter screen name or account ID}

\item{remdr}{an \bold{active} RSelenium \code{\link[RSelenium]{remoteDriver}} object 
(check \code{remdr$getStatus()} to see if the driver is running.)}

\item{since.date}{create date of oldest tweets to get 
Only accepts dates in format '\%Y-\%m-\%d' (Year-month-day: 'YYYY-mm-dd')}

\item{until.date}{create date of most recent (youngest) tweets to get
Only accepts dates in format '\%Y-\%m-\%d' (Year-month-day: 'YYYY-mm-dd')}

\item{date.interval}{date interval passed to 'by' argument of \code{\link[base]{seq.Date}}. Defaults to 'month'.}

\item{write.out}{logical. write out tweet IDs as JSON to disk?
If \code{TRUE} (the default), JSON file will be written 
to path \code{write.out.path} and named \code{write.out.name}.
If \code{FALSE}, \code{write.out.path} and \code{write.out.name} will be ignored.}

\item{write.out.path}{Write out path (directory where to write scraped IDs file)
Will be ignored if \code{write.out = FALSE}}

\item{write.out.name}{JSON file name.
Defaults to 'tw_user_<\code{tw.account}>_tweet_ids_<\code{since.date}>_to_<\code{until.date}>.json' 
Will be ignored if \code{write.out = FALSE}}

\item{sleep}{Seconds to pause between date ranges when iterating over date intervals defined by 
\code{since.date}, \code{until.date} and \code{date.interval}.
Defautls to .5}

\item{verbose}{logical. Print out status messages?}
}
\value{
A \code{\link[tibble]{tibble}} data frame.
     The data frame is empty if an error occurs or no tweet IDs were scraped in the given time range.
     Otherwise it has columns 'account' (<chr>), 'since' (<date>), 'until' (<date>) and 'tweet_id' (<chr>), 
     and one row is one tweet.
}
\description{
Given a twitter account screen name or ID, and start and end dates, 
     function screen-scrapes IDs of historical tweets in time range and returns them
     in a data frame.
     Optionally, the scraped IDs can additionally be written to disk (if \code{write.out = TRUE}).
}
\section{WARNING}{

\itemize{
    \item Function presuposses an active remote Selenium driver.     
    \item Function only accepts dates in format '\%Y-\%m-\%d' (Year-month-day: 'YYYY-mm-dd')     
}
}

